# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/v1/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "summary": "Get current user profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_updateMe",
        "summary": "Update user profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile updated"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/habits": {
      "get": {
        "operationId": "HabitsController_list",
        "summary": "List user habits",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Habits retrieved"
          }
        },
        "tags": [
          "Habits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "operationId": "HabitsController_create",
        "summary": "Create new habit",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Habit created"
          }
        },
        "tags": [
          "Habits"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/habits/{id}/tick": {
      "post": {
        "operationId": "HabitsController_tick",
        "summary": "Mark habit as completed (idempotent)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Habit ticked"
          }
        },
        "tags": [
          "Habits"
        ],
        "security": [
          {
            "IdempotencyKey": []
          },
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/chat": {
      "post": {
        "operationId": "ChatController_chat",
        "summary": "Send message to Drill Sergeant AI",
        "parameters": [],
        "responses": {
          "200": {
            "description": "AI response with structured actions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "reply": {
                      "type": "string"
                    },
                    "updates": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "suggested_actions": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "confidence": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Chat"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "DrillSergeant API",
    "description": "Production backend for DrillSergeant mobile app",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "IdempotencyKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Idempotency-Key"
      }
    },
...
  }
}